+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * [x] Arrange View: Prevent switching between tabs when editing media - #266
 * [x] Arrange View: Selected media still visible after assign media was clicked - #276
 * [x] Arrange View: Changing a media from position full screen to fixed - didn't added a size border- #253
+* [x] Arrange View: Moving Items out of screen-area - #250
 * [x] Media Creation: The Image Preview wasn't resizing on change - #238
 * [x] Media Creation: Visible screen time - minus values were possible - #239